requirements.txt
boto3==1.18.15
botocore==1.21.15
jinja2==3.0.2
numpy==1.21.1
pandas==1.3.1
PyYAML==5.4.1
pyspark==3.1.1
pytz==2021.3
paramiko==2.8.1
requests==2.26.0
retrying==1.3.3
presto-python-client==0.8.2